+++ docs/release_notes.rst
@@ -2,6 +2,13 @@
 Release Notes
 =============
 
+0.4.0
+-----
+
+* Upgraded Javet to v3.0.3
+* Added module ``javet``
+* Added ``JavetReflectionObjectFactory``
+
 0.3.0
 -----